+++ docs/changelog.md
@@ -15,6 +15,12 @@ Types of changes
 ([GH-000](https://github.com/martin-ueding/geo-activity-playground/issues/000))
 -->
 
+## Unreleased
+
+Added:
+
+- Add delete feature for cluster bookmarks. ([GH-357](https://github.com/martin-ueding/geo-activity-playground/issues/357))
+
 ## Version 1.11.0 — 2025-10-13
 
 Added: